Separate out access lists tests
- Add TestAccessLists.py.
- Related to #2097.
